Update TextSolver.java
Zone array is now a static data field of the TextSolver class. This allows for more abstraction methods to help the main method.
Field array is now initialized using array constants instead of using for loops in the main method. This allows the zone array to be created as a static data field.
Zone numbers for Cells are now set in the Cell constructor instead of using for loops in the main method.
Update for loops to enhanced for loops wherever possible to improve readability.
Delete unnecessary test puzzle.
Use Cell methods instead of for loops wherever possible.