Update SimpleSolver.java (first working version)
puzzleSolved and impossible are now created inside the main method instead of outside the main method as static variables. This was done to make testing and debugging easier.
Finished writing solution algorithm. The program now functions properly.
Removed test puzzle. It is no longer needed because any puzzle can be a test puzzle!